+++ test_app/tests/rbac/test_signals.py
@@ -3,7 +3,7 @@ from unittest.mock import MagicMock
 import pytest
 
 from ansible_base.rbac.pipeline import bulk_give_permissions, bulk_remove_permissions
-from ansible_base.rbac.signals import dab_rbac_assignments_created, dab_rbac_assignments_pre_delete
+from ansible_base.rbac.triggers import dab_rbac_assignments_created, dab_rbac_assignments_pre_delete
 from test_app.models import Inventory, User
